Add files via upload
The 'Optimization_NGA' is a binary file, which is pre-compiled on a 64-bit machine.
If you fail to run 'Optimization_NGA', you can use this source code to compile it on your own machine. 

(For this part of the source code, you can refer to Koper, Keith D., Michael E. Wysession, and Douglas A. Wiens. "Multimodal function optimization with a niching genetic algorithm: A seismological example." Bulletin of the Seismological Society of America 89.4 (1999): 978-988)